requirements.txt
flask
flask-socketio
google-generativeai
edge-tts
python-dotenv